+++ constants/constants.go
@@ -1,7 +1,7 @@
 package constants
 
 const (
-	APP_VERSION = `2.3.0`
+	APP_VERSION = `2.4.0`
 
 	ETS = `Euro Truck Simulator 2`
 	ATS = `American Truck Simulator`